feat: prevent logsink if loki enabled
